+++ Extension/src/pages/filtering-log/components/RequestWizard/ruleCreators.js
@@ -20,11 +20,14 @@ import {
     SimpleRegex,
     CosmeticRuleMarker,
     NetworkRule,
+    NETWORK_RULE_OPTIONS,
+    OPTIONS_DELIMITER,
 } from '@adguard/tsurlfilter';
 
 import { strings } from '../../../../common/strings';
 import { Log } from '../../../../common/log';
 
+import { COMMA_DELIMITER } from './constants';
 import { UrlUtils } from './utils';
 
 /**
@@ -228,6 +231,7 @@ export const createExceptionCookieRules = (event) => {
     return patterns;
 };
 
+// TODO: these could be refactored into one createExceptionAdvancedModifierRules
 export const createExceptionRemoveParamRules = (event) => {
     const { frameDomain, requestRule } = event;
 
@@ -243,6 +247,15 @@ export const createExceptionRemoveHeaderRules = (event) => {
     return [
         getUnblockDomainRule(frameDomain, `${NetworkRule.OPTIONS.REMOVEHEADER}=${requestRule.modifierValue}`),
         getUnblockDomainRule(frameDomain, NetworkRule.OPTIONS.REMOVEHEADER),
+    ];
+};
+
+export const createExceptionCspRules = (event) => {
+    const { frameDomain, requestRule } = event;
+
+    return [
+        getUnblockDomainRule(frameDomain, `${NetworkRule.OPTIONS.CSP}=${requestRule.modifierValue}`),
+        getUnblockDomainRule(frameDomain, NetworkRule.OPTIONS.CSP),
     ];
 };
 
@@ -267,4 +280,116 @@ export const createBlockingCookieRule = (event) => {
     patterns.push(blockingRule);
 
     return patterns;
+};
+
+export const createRuleFromParams = ({
+    urlPattern,
+    urlDomain,
+    thirdParty,
+    important,
+    removeParam,
+}) => {
+    let ruleText = urlPattern;
+
+    const options = [];
+
+    // add domain option
+    if (urlDomain) {
+        options.push(`${NETWORK_RULE_OPTIONS.DOMAIN}=${urlDomain}`);
+    }
+    // add important option
+    if (important) {
+        options.push(NETWORK_RULE_OPTIONS.IMPORTANT);
+    }
+    // add third party option
+    if (thirdParty) {
+        options.push(NETWORK_RULE_OPTIONS.THIRD_PARTY);
+    }
+    // add removeparam option
+    if (removeParam) {
+        options.push(NETWORK_RULE_OPTIONS.REMOVEPARAM);
+    }
+
+    if (options.length > 0) {
+        // Pick correct symbol to append options with
+        const hasOptions = ruleText.includes(OPTIONS_DELIMITER);
+        const prefix = hasOptions
+            ? COMMA_DELIMITER
+            : OPTIONS_DELIMITER;
+        ruleText += prefix + options.join(COMMA_DELIMITER);
+    }
+
+    return ruleText;
+};
+
+export const createCssRuleFromParams = (urlPattern, permitDomain) => {
+    let ruleText = urlPattern;
+    if (!permitDomain) {
+        ruleText = ruleText.slice(ruleText.indexOf('#'));
+    }
+
+    return ruleText;
+};
+
+export const createCookieRuleFromParams = ({
+    rulePattern,
+    thirdParty,
+    important,
+}) => {
+    let ruleText = rulePattern;
+
+    const options = [];
+
+    // add important option
+    if (important) {
+        options.push(NETWORK_RULE_OPTIONS.IMPORTANT);
+    }
+    // add third party option
+    if (thirdParty) {
+        options.push(NETWORK_RULE_OPTIONS.THIRD_PARTY);
+    }
+    if (options.length > 0) {
+        ruleText += COMMA_DELIMITER + options.join(COMMA_DELIMITER);
+    }
+
+    return ruleText;
+};
+
+export const getRuleText = (selectedEvent, rulePattern, ruleOptions) => {
+    const {
+        ruleDomain,
+        ruleImportant,
+        ruleThirdParty,
+        ruleRemoveParam,
+    } = ruleOptions;
+
+    const permitDomain = !ruleDomain.checked;
+    const important = !!ruleImportant.checked;
+    const thirdParty = !!ruleThirdParty.checked;
+    const removeParam = !!ruleRemoveParam.checked;
+
+    const domain = permitDomain ? selectedEvent.frameDomain : null;
+
+    let ruleText;
+    if (selectedEvent.element) {
+        ruleText = createCssRuleFromParams(rulePattern, permitDomain);
+    } else if (selectedEvent.cookieName) {
+        ruleText = createCookieRuleFromParams({
+            rulePattern,
+            thirdParty,
+            important,
+        });
+    } else if (selectedEvent.script || selectedEvent?.requestRule?.documentLevelRule) {
+        ruleText = createRuleFromParams({ urlPattern: rulePattern });
+    } else {
+        ruleText = createRuleFromParams({
+            urlPattern: rulePattern,
+            urlDomain: domain,
+            thirdParty,
+            important,
+            removeParam,
+        });
+    }
+
+    return ruleText;
 };