+++ Extension/src/pages/filtering-log/stores/WizardStore.js
@@ -23,7 +23,7 @@ import {
     makeObservable,
 } from 'mobx';
 
-import { NETWORK_RULE_OPTIONS, OPTIONS_DELIMITER } from '@adguard/tsurlfilter';
+import { ContentType } from '@adguard/tswebextension';
 
 import { RULE_OPTIONS } from '../components/RequestWizard/constants';
 import {
@@ -32,9 +32,11 @@ import {
     createExceptionCssRule,
     createExceptionRemoveParamRules,
     createExceptionRemoveHeaderRules,
+    createExceptionCspRules,
     createExceptionScriptRule,
     createBlockingCookieRule,
     splitToPatterns,
+    getRuleText,
 } from '../components/RequestWizard/ruleCreators';
 import { messenger } from '../../services/messenger';
 
@@ -50,13 +52,12 @@ export const ADDED_RULE_STATES = {
     UNBLOCK: 'unblock',
 };
 
-const MODIFIERS_DELIMITER = ',';
-
 const defaultRuleOptions = {
     [RULE_OPTIONS.RULE_DOMAIN]: { checked: false },
     [RULE_OPTIONS.RULE_THIRD_PARTY]: { checked: false },
     [RULE_OPTIONS.RULE_IMPORTANT]: { checked: false },
 };
+
 class WizardStore {
     constructor(rootStore) {
         this.rootStore = rootStore;
@@ -202,138 +203,13 @@ class WizardStore {
         this.ruleText = ruleText;
     }
 
-    createRuleFromParams = ({
-        urlPattern,
-        urlDomain,
-        thirdParty,
-        important,
-        mandatoryOptions,
-        removeParam,
-    }) => {
-        let ruleText = urlPattern;
-
-        let options = [];
-
-        // add domain option
-        if (urlDomain) {
-            options.push(`${NETWORK_RULE_OPTIONS.DOMAIN}=${urlDomain}`);
-        }
-        // add important option
-        if (important) {
-            options.push(NETWORK_RULE_OPTIONS.IMPORTANT);
-        }
-        // add third party option
-        if (thirdParty) {
-            options.push(NETWORK_RULE_OPTIONS.THIRD_PARTY);
-        }
-        // add removeparam option
-        if (removeParam) {
-            options.push(NETWORK_RULE_OPTIONS.REMOVEPARAM);
-        }
-        if (mandatoryOptions) {
-            options = options.concat(mandatoryOptions);
-        }
-        if (options.length > 0) {
-            ruleText += OPTIONS_DELIMITER + options.join(MODIFIERS_DELIMITER);
-        }
-
-        return ruleText;
-    };
-
-    createCssRuleFromParams = (urlPattern, permitDomain) => {
-        let ruleText = urlPattern;
-        if (!permitDomain) {
-            ruleText = ruleText.slice(ruleText.indexOf('#'));
-        }
-
-        return ruleText;
-    };
-
-    createCookieRuleFromParams = ({
-        rulePattern,
-        thirdParty,
-        important,
-    }) => {
-        let ruleText = rulePattern;
-
-        const options = [];
-
-        // add important option
-        if (important) {
-            options.push(NETWORK_RULE_OPTIONS.IMPORTANT);
-        }
-        // add third party option
-        if (thirdParty) {
-            options.push(NETWORK_RULE_OPTIONS.THIRD_PARTY);
-        }
-        if (options.length > 0) {
-            ruleText += MODIFIERS_DELIMITER + options.join(MODIFIERS_DELIMITER);
-        }
-
-        return ruleText;
-    };
-
-    getRuleText(selectedEvent, rulePattern, ruleOptions) {
-        // if rule was edited by user return it as is
-        if (this.ruleText !== null) {
-            return this.ruleText;
-        }
-
-        const {
-            ruleDomain,
-            ruleImportant,
-            ruleThirdParty,
-            ruleRemoveParam,
-        } = ruleOptions;
-
-        const permitDomain = !ruleDomain.checked;
-        const important = !!ruleImportant.checked;
-        const thirdParty = !!ruleThirdParty.checked;
-        const removeParam = !!ruleRemoveParam.checked;
-
-        const domain = permitDomain ? selectedEvent.frameDomain : null;
-
-        let mandatoryOptions = null;
-
-        // Deal with csp rule
-        const { requestRule } = selectedEvent;
-        if (requestRule && requestRule.cspRule) {
-            mandatoryOptions = [NETWORK_RULE_OPTIONS.CSP];
-        }
-
-        if (selectedEvent.replaceRules) {
-            mandatoryOptions = [NETWORK_RULE_OPTIONS.REPLACE];
-        }
-
-        let ruleText;
-        if (selectedEvent.element) {
-            ruleText = this.createCssRuleFromParams(rulePattern, permitDomain);
-        } else if (selectedEvent.cookieName) {
-            ruleText = this.createCookieRuleFromParams({
-                rulePattern,
-                thirdParty,
-                important,
-            });
-        } else if (selectedEvent.script || selectedEvent?.requestRule?.documentLevelRule) {
-            ruleText = this.createRuleFromParams({ urlPattern: rulePattern });
-        } else {
-            ruleText = this.createRuleFromParams({
-                urlPattern: rulePattern,
-                urlDomain: domain,
-                thirdParty,
-                important,
-                mandatoryOptions,
-                removeParam,
-            });
-        }
-
-        return ruleText;
-    }
-
     @computed
     get rule() {
         const { logStore } = this.rootStore;
-        return this.getRuleText(logStore.selectedEvent, this.rulePattern, this.ruleOptions);
+        // If the rule was edited by the user, it will be returned as is
+        return this.ruleText === null
+            ? getRuleText(logStore.selectedEvent, this.rulePattern, this.ruleOptions)
+            : this.ruleText;
     }
 
     @computed
@@ -367,6 +243,10 @@ class WizardStore {
 
             if (selectedEvent.removeHeader) {
                 patterns = createExceptionRemoveHeaderRules(selectedEvent);
+            }
+
+            if (selectedEvent.requestType === ContentType.Csp) {
+                patterns = createExceptionCspRules(selectedEvent);
             }
 
             this.setRulePattern(patterns[0]);